formatters refactoring
